:electron: Add example script QM2MFAST
Add an example script that provides an overview of how to prepare the
inputs for MFAST from the outputs of a QuakeMigrate run.

The user will need to edit the information within the highlighted
section. It is also here that they can perform any pre-filtering of the
event catalogue, or this can be performed separately then passed into
the translator.